Dev To Main (#3)
* Implement a task structure (#1)

* Add Task structure

Add a structure for Task and implement some methods. One method is a print
method for Task. The other method searches a DB for the task by ID.

* Add option to get info on a task given a task ID

* Implement Category structure and more database fixes (#2)

* Add Task structure

Add a structure for Task and implement some methods. One method is a print
method for Task. The other method searches a DB for the task by ID.

* Add option to get info on a task given a task ID

* Implement a category structure

Implement a print function for a function

Implement returning a vector of Category instead of a vector of Strings

* Inserting in default values into a newly created database

* Implement printing all categories using a vector of categories inste

of strings. Also panic with an error message instead of panicing without a message