Implement more structures (#4)
* Add Task structure

Add a structure for Task and implement some methods. One method is a print
method for Task. The other method searches a DB for the task by ID.

* Add option to get info on a task given a task ID

* Implement a category structure

Implement a print function for a function

Implement returning a vector of Category instead of a vector of Strings

* Inserting in default values into a newly created database

* Implement printing all categories using a vector of categories inste

of strings. Also panic with an error message instead of panicing without a message

* Format fixing

* Add a get all status function

* Fixes for syntax not picked up when testing zee text editor

* No longer need testing file

* Simplify SQL stataments for creating database

* Call self.name instead of self.id to print correct value